+++ slack_sdk/audit_logs/__init__.py
@@ -1,0 +1,2 @@
+from .v1.client import AuditLogsClient  # noqa
+from .v1.response import AuditLogsResponse  # noqa
